python/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
mariadb==1.0.5
MarkupSafe==1.1.1
python-redis-lock==3.7.0
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
selenium==3.141.0
urllib3==1.26.2
Werkzeug==1.0.1
